Move all react components to componenets/ folder
