+++ src/components/Details/View.tsx
