+++ src/components/Rating/index.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react'
 import styled from '@emotion/styled/macro'
 
 import theme from 'theme'
-import * as Skelet from 'Skeleton'
+import * as Skelet from 'components/Skeleton'
 import * as api from 'api'
 
 const limitRating = (rating: number): string =>
